Refactored DB module
